+++ lib/data/datasource/remote/user/user_remote.dart
@@ -1,8 +1,8 @@
 import '../../../models/address_model.dart';
-import '../../../models/requests/add_address_request.dart';
-import '../../../models/requests/change_password_request.dart';
-import '../../../models/requests/update_address_request.dart';
-import '../../../models/requests/update_user_profile_request.dart';
+import '../../../models/request/add_address_request.dart';
+import '../../../models/request/change_password_request.dart';
+import '../../../models/request/update_address_request.dart';
+import '../../../models/request/update_user_profile_request.dart';
 import '../../../models/user_model.dart';
 
 abstract class UserRemoteDataSource {
