+++ lib/data/models/request/update_user_profile_request.dart
@@ -3,6 +3,8 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 part 'update_user_profile_request.freezed.dart';
 part 'update_user_profile_request.g.dart';
 
+
+
 @freezed
 class UpdateUserProfileRequest with _$UpdateUserProfileRequest {
   const factory UpdateUserProfileRequest({
