+++ lib/data/repositories/user_repository_impl.dart
@@ -7,10 +7,10 @@ import '../../domain/entities/user.dart';
 import '../../domain/repositories/user_repository.dart';
 import '../core/utils/repository_call_handler.dart';
 import '../datasource/remote/user/user_remote.dart';
-import '../models/requests/update_user_profile_request.dart';
-import '../models/requests/change_password_request.dart';
-import '../models/requests/add_address_request.dart';
-import '../models/requests/update_address_request.dart';
+import '../models/request/update_user_profile_request.dart';
+import '../models/request/change_password_request.dart';
+import '../models/request/add_address_request.dart';
+import '../models/request/update_address_request.dart';
 
 class UserRepositoryImpl implements UserRepository {
   final UserRemoteDataSource _remoteDataSource;
